Module type Pbkdf.S

RFC 2898 specifies two password-based key derivation functions (PBKDF1 and PBKDF2), which are abstracted over a specific hash/pseudorandom function.

val pbkdf1 : password:string -> salt:string -> count:int -> dk_len:int -> string

pbkdf1 password salt count dk_len is dk, the derived key of dk_len octets. The salt must be eight octets, count the iteration count.

raises Invalid_argument
when either salt is not eight octets long or either count or dk_len are not valid.

val pbkdf2 : password:string -> salt:string -> count:int -> dk_len:int32 -> string

pbkdf2 password salt count dk_len is dk, the derived key of dk_len octets.

raises Invalid_argument
when either count or dk_len are not valid